Score Update
Added the functionality for when match streak that are a multiple of 5
happens up to a max of 30 gives the players a bonus to their score.

The Match areas on the wheel were changes to polygon colliders from edge
colliders to better detect collision.

Tiles have hade their box colliders made smaller.

Commented code in multiple scripts.

Issue with circular timer was resolved by moving the move tile call in
the circular timer script to the reset function and disabling a
timeManger object in the scene (Might need to delete that).