Sounds and Difficulty variables
Added the sound FX and theme music to the main game and slot game.

Added three other tiles to increase the difficulty of the game. When a
unmatch tile is active the player must do noting to earn points.

If a match tile is active then doing nothing will decrease the time bar.