Game over overhaul and comments
Created Prefab of the score UI and over hauled the Game over and Stat
panel UI's. Commented some code for clarity and removed the
functionality where if the player does nothing on an unmatch tile the
score will not go up. If the player does noting on a tile they need to
match they then the streak is lost and slider goes down.