Assets Uploaded
Put color variants for the match wheel, button template and upcoming
items in. Included a note with the dimensions of the buttons and
upcoming items, but I did divide it up myself. Added the Backgrounds.
Included the Slot machine along with 3 different "bars" to mark which
what symbols need to match.